chore: Extract xml queries
