+++ editors/base-element-editor.ts
@@ -14,6 +14,11 @@ import {
   identity,
 } from '@openenergytools/scl-lib';
 import '@openenergytools/filterable-lists/dist/action-list.js';
+import {
+  isFCDACompatibleWithIED,
+  queryLDevice,
+  queryLN,
+} from '../foundation/utils/xml.js';
 
 // eslint-disable-next-line no-shadow
 export enum ControlBlockCopyStatus {
@@ -78,54 +83,26 @@ export class BaseElementEditor extends ScopedElementsMixin(LitElement) {
   }
 
   // eslint-disable-next-line class-methods-use-this
-  protected queryLDevice(ied: Element, inst: string): Element | null {
-    return ied.querySelector(
-      `:scope > AccessPoint > Server > LDevice[inst="${inst}"]`
-    );
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  protected buildLnQuery(
-    lnClass: string,
-    inst: string,
-    prefix: string | null
-  ): string {
-    const ln0Query = `:scope > LN0[lnClass="${lnClass}"]`;
-    let lnQuery = `:scope > LN[lnClass="${lnClass}"][inst="${inst}"]`;
-
-    if (prefix) {
-      lnQuery += `[prefix="${prefix}"]`;
-    } else {
-      lnQuery += `:not([prefix]), ${lnQuery}[prefix=""]`;
-    }
-
-    return `${ln0Query}, ${lnQuery}`;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  protected buildLnQueryForControl(control: Element): string {
+  protected queryLnForControl(ied: Element, control: Element): Element | null {
     const lDevice = control.closest('LDevice');
-    const lnOrLn0 = control.closest('LN0') ?? control.closest('LN');
+    const lnOrLn0 = control.closest('LN0, LN');
 
-    if (!lnOrLn0) {
-      throw new Error('ControlBlock must be a child of LN or LN0');
+    if (!lnOrLn0 || !lDevice) {
+      throw new Error('ControlBlock must be a child of LN or LN0 and LDevice');
     }
 
-    let lnQuery = '';
+    const ldInst = lDevice.getAttribute('inst') ?? '';
+    const lDeviceInIed = queryLDevice(ied, ldInst);
 
-    if (lnOrLn0.tagName === 'LN0') {
-      lnQuery = 'LN0';
-    } else {
-      const lnClass = lnOrLn0.getAttribute('lnClass');
-      // TODO: Handle optional prefix
-      const prefix = lnOrLn0.getAttribute('prefix');
-      const inst = lnOrLn0.getAttribute('inst');
-      lnQuery = `LN[lnClass="${lnClass}"][inst="${inst}"][prefix="${prefix}"]`;
+    if (!lDeviceInIed) {
+      return null;
     }
 
-    return `:scope > AccessPoint > Server > LDevice[inst="${lDevice?.getAttribute(
-      'inst'
-    )}"] > ${lnQuery}`;
+    const lnClass = lnOrLn0.getAttribute('lnClass') ?? '';
+    const inst = lnOrLn0.getAttribute('inst') ?? '';
+    const prefix = lnOrLn0.getAttribute('prefix');
+
+    return queryLN(lDeviceInIed, lnClass, inst, prefix);
   }
 
   // eslint-disable-next-line class-methods-use-this
@@ -138,138 +115,11 @@ export class BaseElementEditor extends ScopedElementsMixin(LitElement) {
   }
 
   // eslint-disable-next-line class-methods-use-this
-  protected hasDataType(
-    dataTypeTemplates: Element,
-    lnType: string,
-    doSegments: string[],
-    daSegments: string[]
-  ): boolean {
-    const lNodeType = dataTypeTemplates.querySelector(
-      `LNodeType[id="${lnType}"]`
-    );
-    if (!lNodeType) {
-      return false;
-    }
-
-    let currentElement = lNodeType;
-
-    for (const doSegement of doSegments) {
-      const dataObject = currentElement.querySelector(
-        `DO[name="${doSegement}"], SDO[name="${doSegement}"]`
-      );
-      if (!dataObject) {
-        return false;
-      }
-
-      // console.log('DO', dataObject);
-
-      const doType = dataObject.getAttribute('type');
-      if (!doType) {
-        return false;
-      }
-
-      const doTypeElement = dataTypeTemplates.querySelector(
-        `DOType[id="${doType}"]`
-      );
-      if (!doTypeElement) {
-        return false;
-      }
-
-      // console.log('DOType', doTypeElement);
-
-      currentElement = doTypeElement;
-    }
-
-    for (const [index, daSegment] of daSegments.entries()) {
-      // TODO: Handle enum types
-      const dataAttribute = currentElement.querySelector(
-        `DA[name="${daSegment}"], BDA[name="${daSegment}"]`
-      );
-      if (!dataAttribute) {
-        return false;
-      }
-
-      // console.log('DA', dataAttribute);
-
-      if (index === daSegments.length - 1) {
-        // Do not search for type, because last entry is a leaf node and should be basic type
-        break;
-      }
-
-      const daType = dataAttribute.getAttribute('type');
-      if (!daType) {
-        return false;
-      }
-
-      const daTypeElement = dataTypeTemplates.querySelector(
-        `DAType[id="${daType}"]`
-      );
-      if (!daTypeElement) {
-        return false;
-      }
-
-      // console.log('DAType', daTypeElement);
-
-      currentElement = daTypeElement;
-    }
-
-    return true;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  protected isFCDACompatibleWithIED(fcda: Element, ied: Element): boolean {
-    const ldInst = fcda.getAttribute('ldInst');
-    const lnClass = fcda.getAttribute('lnClass');
-    const prefix = fcda.getAttribute('prefix');
-    const lnInst = fcda.getAttribute('lnInst');
-    const doName = fcda.getAttribute('doName');
-    const daName = fcda.getAttribute('daName');
-
-    const daSegments = daName ? daName.split('.') : [];
-
-    if (!ldInst || !lnClass || !doName || !lnInst) {
-      return false;
-    }
-
-    const doSegments = doName.split('.');
-
-    if (doSegments.length === 0) {
-      return false;
-    }
-
-    const lDevice = this.queryLDevice(ied, ldInst);
-    if (!lDevice) {
-      return false;
-    }
-
-    const lnQuery = this.buildLnQuery(lnClass, lnInst, prefix);
-    const ln = lDevice.querySelector(lnQuery);
-    if (!ln) {
-      return false;
-    }
-
-    const lnType = ln.getAttribute('lnType');
-    if (!lnType) {
-      return false;
-    }
-
-    const dataTypeTemplates =
-      ied.parentElement?.querySelector('DataTypeTemplates');
-    if (!dataTypeTemplates) {
-      return false;
-    }
-
-    // console.log(`Checking ${ied.getAttribute('name')} for ${lnType}, ${doSegments}, ${daSegments}`);
-    return this.hasDataType(dataTypeTemplates, lnType, doSegments, daSegments);
-  }
-
-  // eslint-disable-next-line class-methods-use-this
   protected getCopyControlBlockCopyStatus(
     controlBlock: Element,
     otherIED: Element
   ): ControlBlockCopyStatus {
-    const lnQuery = this.buildLnQueryForControl(controlBlock);
-    const ln = otherIED.querySelector(lnQuery);
+    const ln = this.queryLnForControl(otherIED, controlBlock);
 
     if (!ln) {
       return ControlBlockCopyStatus.IEDStructureIncompatible;
@@ -298,7 +148,7 @@ export class BaseElementEditor extends ScopedElementsMixin(LitElement) {
 
     const fcdas = Array.from(dataSet.querySelectorAll('FCDA'));
     for (const fcda of fcdas) {
-      const isCompatible = this.isFCDACompatibleWithIED(fcda, otherIED);
+      const isCompatible = isFCDACompatibleWithIED(fcda, otherIED);
 
       if (!isCompatible) {
         // console.log(`FCDA is not compatible`, fcda);
@@ -320,8 +170,7 @@ export class BaseElementEditor extends ScopedElementsMixin(LitElement) {
     }
 
     const inserts = selectedOptions.flatMap(o => {
-      const lnQuery = this.buildLnQueryForControl(o.control);
-      const ln = o.ied.querySelector(lnQuery);
+      const ln = this.queryLnForControl(o.ied, o.control);
       const dataSet = this.getDataSet(o.control);
 
       if (!ln || !dataSet) {
